Added Grid Sizing button and functionality
Added button at top of page to the HTML, this allows you to click it
and resize the grid to be more or less pixelated. Added a function that
ensures input is between 1 and 100, as grid sizes greater than 100x100
will reduce performance of the app. Proud of this update. Fist pumping
on a plane over Mexico flying from Leon to Merida.