Added shading functionality and minor CSS changes
Reworked changeColor function to marginally increase opacity of squares
rather than changing it completely to gray on the first user
interaction. This allows for the user to have some room for error, and
also allows them to incorporate shading into their drawings. I also
centered the container and button, which I felt looked better. This
wraps up the etchASketch project. On to the next one!